# packages/shared-types/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/intent": {
      "post": {
        "operationId": "IntentController_dispatch",
        "summary": "Dispatch intent to AI service",
        "description": "Returns request_id for SSE stream pickup via GET /intent/stream?id=<rid>. S-04 T03: mode field selects Variant B (ai_augmented) or Variant A (basic_fallback) per D-S04-03 LAW Adaptive Single Endpoint. Sx05-3-CODE HOTFIX (D-S05-13 LAW): @UseGuards(JwtAuthGuard) added — req.user.id forwarded to AI as PostIntentBody.user_id so cart_by_text graph operates on correct authenticated cart.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentRequestDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "intent"
        ],
        "security": [
          {
            "icp_session": []
          }
        ]
      }
    },
...
    "/api/v1/intent/{rid}/action": {
      "post": {
        "operationId": "IntentActionController_action",
        "summary": "Resume interrupted intent graph with user choice",
        "description": "Forwards action body to AI service /intent/{rid}/resume endpoint. AI calls graph.astream(Command(resume=<choice>)) to continue from Pattern P2 interrupt checkpoint. SSE events flow back via existing GET /intent/stream connection (Option Z Redis pub/sub).",
        "parameters": [
          {
            "name": "rid",
            "required": true,
            "in": "path",
            "description": "request_id from POST /intent response",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentActionDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "intent"
        ],
        "security": [
          {
            "icp_session": []
          }
        ]
      }
    },
...
    "/api/v1/auth/landing": {
      "get": {
        "operationId": "LandingController_landing",
        "summary": "Resolve landing redirect (last shop hoặc onboarding)",
        "description": "ADR-046 amend (c): đọc sessions.last_active_tenant_id → redirect_url `/s/<slug>` (last_active) hoặc `/onboarding` (chưa có hint). Dùng cho root URL.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LandingResponseDto"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "icp_session": []
          }
        ]
      }
    },
    "/api/v1/auth/tenants": {
      "get": {
        "operationId": "LandingController_myTenants",
        "summary": "Danh sách shop user là member (onboarding/switcher)",
        "description": "ADR-046 amend (c): FE không đọc được tenant_ids (httpOnly JWT) → endpoint join slug/name. Dùng cho /onboarding (chọn shop) + tenant-switcher.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyTenantsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "icp_session": []
          }
        ]
      }
    },
...
    "/api/v1/dashboard/stats": {
      "get": {
        "operationId": "DashboardController_getStats",
        "summary": "Get dashboard KPI stats — stub per S-03 D-10",
        "description": "Returns 200 with hardcoded JSON `{orders_today, revenue_today, inventory_count, currency: \"VND\"}` matching mockup StatBar text values. Requires icp_session cookie. Real DB aggregations defer to future slice when S-05 Cart/Order ships data. Per S-03 D-10 MAR-1 Q5 RESOLVED + DM-14.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStatsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid icp_session cookie"
          }
        },
        "tags": [
          "dashboard"
        ],
        "security": [
          {
            "icp_session": []
          }
        ]
      }
    },
    "/api/v1/dashboard/insight": {
      "get": {
        "operationId": "DashboardController_getInsight",
        "summary": "Get home hero AI insight — REAL analytics (S-10 detect_anomaly)",
        "description": "Returns `{delta_pct, direction, cause_count, has_data}` from analytics.detect_anomaly (merchant-wide 7d-vs-prior-7d + count of flagged categories). Replaces the hard-coded \"giảm 12% / 2 nguyên nhân\" in HeroInsightCard. Same engine as /intent-07 so numbers agree. Requires icp_session cookie.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardInsightDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid icp_session cookie"
          }
        },
        "tags": [
          "dashboard"
        ],
        "security": [
          {
            "icp_session": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "ICP Gateway API",
    "description": "Intelligent Commerce Platform — REST API.\n\n4 mutating endpoints require `Idempotency-Key` header (UUID v4) per docs/03_API_CONTRACTS.md §1.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "icp_session",
        "description": "httpOnly session cookie issued by POST /auth/login (per ADR-019 + S-03 C-01)."
      }
    },
    "schemas": {
...
      "IntentRequestDto": {
        "type": "object",
        "properties": {
          "modality": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "voice"
            ]
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000000
          },
          "hint": {
            "type": "string",
            "enum": [
              "import",
              "buy",
              "search",
              "recommend",
              "cart_clear_confirm",
              "cart_view_with_stock_check",
              "analyze"
            ]
          },
          "mode": {
            "type": "string",
            "enum": [
              "ai_augmented",
              "basic_fallback"
            ],
            "default": "ai_augmented"
          },
          "text_query": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "filters": {
            "type": "object",
            "properties": {
              "brand": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "category": {
                "type": "string",
                "minLength": 1,
                "maxLength": 50
              },
              "attributes": {
                "type": "object",
                "additionalProperties": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100
                }
              }
            }
          }
        },
        "required": [
          "modality"
        ]
      },
      "IntentActionDto": {
        "type": "object",
        "properties": {
          "card_id": {
            "type": "string"
          },
          "choice": {
            "type": "string",
            "enum": [
              "accept",
              "reject",
              "retry_ai",
              "continue_basic",
              "add_to_cart",
              "skip",
              "confirm_clear",
              "cancel_clear",
              "resolve_remove",
              "resolve_replace",
              "clarify_pick"
            ]
          },
          "value": {
            "type": "object",
            "additionalProperties": {}
          },
          "_meta": {
            "type": "object",
            "properties": {
              "attempt_n": {
                "type": "integer",
                "minimum": 0,
                "exclusiveMinimum": true
              }
            },
            "required": [
              "attempt_n"
            ]
          }
        },
        "required": [
          "choice"
        ]
      },
...
      "LandingResponseDto": {
        "type": "object",
        "properties": {
          "redirect_url": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "last_active",
              "onboarding"
            ]
          }
        },
        "required": [
          "redirect_url",
          "source"
        ]
      },
      "MyTenantsResponseDto": {
        "type": "object",
        "properties": {
          "tenants": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "tenant_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "slug": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              },
              "required": [
                "tenant_id",
                "slug",
                "name"
              ]
            }
          }
        },
        "required": [
          "tenants"
        ]
      },
...
      "DashboardStatsDto": {
        "type": "object",
        "properties": {
          "orders_today": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "revenue_today": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "inventory_count": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "currency": {
            "type": "string",
            "enum": [
              "VND"
            ]
          }
        },
        "required": [
          "orders_today",
          "revenue_today",
          "inventory_count",
          "currency"
        ]
      },
      "DashboardInsightDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}