# packages/shared-types/src/api/models/LandingResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LandingResponseDto = {
  redirect_url: string;
  source: 'last_active' | 'onboarding';
};