# packages/shared-types/src/api/models/MyTenantsResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MyTenantsResponseDto = {
  tenants: Array<{
    tenant_id: string;
    slug: string;
    name: string;
  }>;
};